SUW: Don't make google suw use material_light
The material_light theme causes gms to crash with "Error inflating
class TextView" when attempting to do a device to device restore in
google suw. This change allows it to use the default glif_light.

Also removed the "Partner SUW Strings" comment, as the theme_type
was the only one.

Change-Id: I287f201799aba81c19869000e3aadb47ef67299b